+++ A3A/addons/hals/Addons/store/config.hpp
@@ -80,7 +80,8 @@ class cfgHALsStore
 		#include "config\ffp.hpp"
 		#include "config\ylarms.hpp"
 		#include "config\ProjInfAD.hpp"
-		#include "config\JCAArsenal.hpp"
+		#include "config\JCAIA.hpp"
+		#include "config\JCAIE.hpp"
 		#include "config\hafm.hpp"
 	};
 
@@ -942,7 +943,7 @@ class cfgHALsStore
 				"underbarrelProjInfAD"
 			};
 		};
-		class jca_stock
+		class jcaia_stock
 		{
 			displayName = $STR_ARMS_DEALER_STORE;
 			categories[] = {
@@ -958,6 +959,14 @@ class cfgHALsStore
 				"launchersJCA",
 				"launcherMagazinesJCA",
 				"miscJCA"
+			};
+		};
+		class jcaie_stock
+		{
+			displayName = $STR_ARMS_DEALER_STORE;
+			categories[] = {
+				"vestsJCA",
+				"facewearJCA"
 			};
 		};
 	};
