+++ A3A/addons/hals/Addons/store/config/JCAIE.hpp
@@ -1,0 +1,83 @@
+class vestsJCA
+{
+	displayName = __EVAL(formatText ["%1 %2 %3 %4", localize "STR_A3AU_JCA_ARSENAL", localize "STR_A3AU_vests"]);
+	picture = "a3\ui_f\data\gui\Rsc\RscDisplayArsenal\backpack_ca.paa";
+
+	ITEM(JCA_V_CarrierRigKBT_01_combat_black_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_combat_olive_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_combat_sand_F, 700, MISC_STOCK);
+	
+	ITEM(JCA_V_CarrierRigKBT_01_command_black_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_command_olive_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_command_sand_F, 700, MISC_STOCK);
+
+	ITEM(JCA_V_CarrierRigKBT_01_compact_black_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_compact_olive_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_compact_sand_F, 700, MISC_STOCK);
+
+	ITEM(JCA_V_CarrierRigKBT_01_CQB_black_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_CQB_olive_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_CQB_sand_F, 700, MISC_STOCK);
+
+	ITEM(JCA_V_CarrierRigKBT_01_crew_black_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_crew_olive_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_crew_sand_F, 700, MISC_STOCK);
+
+	ITEM(JCA_V_CarrierRigKBT_01_heavy_black_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_heavy_olive_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_heavy_sand_F, 700, MISC_STOCK);
+
+	ITEM(JCA_V_CarrierRigKBT_01_light_black_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_light_olive_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_light_sand_F, 700, MISC_STOCK);
+
+	ITEM(JCA_V_CarrierRigKBT_01_recon_black_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_recon_olive_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_recon_sand_F, 700, MISC_STOCK);
+
+	ITEM(JCA_V_CarrierRigKBT_01_tactical_black_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_tactical_olive_F, 700, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_tactical_sand_F, 700, MISC_STOCK);
+
+	ITEM(JCA_V_CarrierRigKBT_01_black_F, 600, MISC_STOCK);//Same Armor Values, Just lower load
+	ITEM(JCA_V_CarrierRigKBT_01_olive_F, 600, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_sand_F, 600, MISC_STOCK);
+
+	ITEM(JCA_V_CarrierRigKBT_01_holster_black_F, 100, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_holster_olive_F, 100, MISC_STOCK);
+	ITEM(JCA_V_CarrierRigKBT_01_holster_sand_F, 100, MISC_STOCK);
+};
+
+class facewearJCA
+{
+	displayName = __EVAL(formatText ["%1 %2 %3 %4", localize "STR_A3AU_JCA_ARSENAL", localize "STR_A3AU_facewear"]);
+	picture = "a3\ui_f\data\gui\Rsc\RscDisplayArsenal\backpack_ca.paa";
+
+	ITEM(JCA_G_Headset_Combat_01_clear_black_F, 30, MISC_STOCK);
+	ITEM(JCA_G_Headset_Combat_01_clear_olive_F, 30, MISC_STOCK);
+	ITEM(JCA_G_Headset_Combat_01_clear_sand_F, 30, MISC_STOCK);
+
+	ITEM(JCA_G_AirPurifyingRespirator_03_black_clear_F, 30, MISC_STOCK);
+	ITEM(JCA_G_AirPurifyingRespirator_03_black_tinted_F, 30, MISC_STOCK);
+
+	ITEM(JCA_G_AirPurifyingRespirator_03_olive_clear_F, 30, MISC_STOCK)
+	ITEM(JCA_G_AirPurifyingRespirator_03_olive_tinted_F, 30, MISC_STOCK);
+
+	ITEM(JCA_G_AirPurifyingRespirator_03_sand_clear_F, 30, MISC_STOCK);
+	ITEM(JCA_G_AirPurifyingRespirator_03_sand_tinted_F, 30, MISC_STOCK);
+
+	ITEM(JCA_G_Glasses_Tactical_amber_black_F, 30, MISC_STOCK);
+	ITEM(JCA_G_Glasses_Tactical_clear_black_F, 30, MISC_STOCK);
+	ITEM(JCA_G_Glasses_Tactical_faded_black_F, 30, MISC_STOCK);
+	ITEM(JCA_G_Glasses_Tactical_tinted_black_F, 30, MISC_STOCK);
+
+	ITEM(JCA_G_Glasses_Tactical_amber_olive_F, 30, MISC_STOCK);
+	ITEM(JCA_G_Glasses_Tactical_clear_olive_F, 30, MISC_STOCK);
+	ITEM(JCA_G_Glasses_Tactical_faded_olive_F, 30, MISC_STOCK);
+	ITEM(JCA_G_Glasses_Tactical_tinted_olive_F, 30, MISC_STOCK);
+
+	ITEM(JCA_G_Glasses_Tactical_amber_sand_F, 30, MISC_STOCK);
+	ITEM(JCA_G_Glasses_Tactical_clear_sand_F, 30, MISC_STOCK);
+	ITEM(JCA_G_Glasses_Tactical_faded_sand_F, 30, MISC_STOCK);
+	ITEM(JCA_G_Glasses_Tactical_tinted_sand_F, 30, MISC_STOCK);
+};
